requirements.txt
## pyttsx3
nltk
fastapi
requests
uvicorn
python-multipart
python-dotenv
streamlit